make some things less public and rename them
