+++ rust/sedona-query-planner/src/spatial_join_factory.rs
