+++ rust/sedona-spatial-join/src/factory.rs
@@ -22,8 +22,8 @@ use datafusion::physical_plan::ExecutionPlan;
 use datafusion_common::{JoinSide, Result};
 use datafusion_physical_expr::PhysicalExpr;
 use sedona_common::sedona_internal_err;
-use sedona_query_planner::extension_planner::{PlanSpatialJoinArgs, SpatialJoinFactory};
 use sedona_query_planner::probe_shuffle_exec::ProbeShuffleExec;
+use sedona_query_planner::spatial_join_factory::{PlanSpatialJoinArgs, SpatialJoinFactory};
 use sedona_query_planner::spatial_predicate::{DistancePredicate, KNNPredicate, RelationPredicate};
 use sedona_schema::datatypes::SedonaType;
 use sedona_schema::matchers::ArgMatcher;
